actually, no need to template it
