+++ CMakeLists.txt
@@ -9,7 +9,7 @@ set(LOADABLE_EXTENSION_NAME ${TARGET_NAME}_loadable_extension)
 project(${TARGET_NAME})
 include_directories(src/include)
 
-set(EXTENSION_SOURCES src/variant_extension.cpp src/variant_functions.cpp src/to_variant.cpp src/variant_to_varchar.cpp src/from_variant.cpp src/variant_extract.cpp)
+set(EXTENSION_SOURCES src/variant_extension.cpp src/variant_functions.cpp src/to_variant.cpp src/variant_to_varchar.cpp src/from_variant.cpp src/variant_extract.cpp src/variant_utils.cpp)
 
 build_static_extension(${TARGET_NAME} ${EXTENSION_SOURCES})
 build_loadable_extension(${TARGET_NAME} " " ${EXTENSION_SOURCES})
